+++ src/temp/ProjectMapTabs/SingleTab.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './ProjectViewTabs.module.scss';
+import styles from './ProjectMapTabs.module.scss';
 
 const SingleTab = ({ icon, title, isSelected }) => {
   return (
