+++ src/temp/stories/SingleMapTab.stories.tsx
@@ -1,6 +1,6 @@
 import SatelliteIcon from '../../../public/assets/images/icons/SatelliteIcon';
 import type { Meta, StoryObj } from '@storybook/react';
-import SingleTab from '../ProjectViewTabs/SingleTab';
+import SingleTab from '../ProjectMapTabs/SingleTab';
 
 const meta: Meta<typeof SingleTab> = {
   component: SingleTab,
@@ -9,7 +9,7 @@ const meta: Meta<typeof SingleTab> = {
 export default meta;
 type Story = StoryObj<typeof SingleTab>;
 
-export const TimeTravelSelected: Story = {
+export const Selected: Story = {
   args: {
     icon: <SatelliteIcon color={'#fff'} />,
     title: 'Time Travel',
@@ -17,7 +17,7 @@ export const TimeTravelSelected: Story = {
   },
 };
 
-export const TimeTravelUnSelected: Story = {
+export const Unselected: Story = {
   args: {
     icon: <SatelliteIcon color={'#000'} />,
     title: 'Time Travel',
